fix(AAP-84199): use SynKebabMenu after Nx-to-Syn rename
