+++ frontend/packages/syntara-ui/src/routes/access-management/GroupsTab.tsx
@@ -11,7 +11,6 @@ import { SynConfirmationDialog } from '../../components/dialogs/SynConfirmationD
 import { DisabledWithTooltip } from '../../components/DisabledWithTooltip'
 import { IconLabel } from '../../components/IconLabel'
 import type { KebabAction } from '../../components/SynKebabMenu'
-import { SynKebabMenu } from '../../components/SynKebabMenu'
 import { SynListPanelTable, SynListPanelToolbar, SynListPanelView } from '../../components/panels/list/SynListPanel'
 import { SynEmptyStateNoData } from '../../components/states/SynEmptyStateNoData'
 import { DateCell } from '../../components/table/DateCell'
