+++ frontend/packages/syntara-ui/src/routes/access-management/ProjectsTab.tsx
@@ -10,7 +10,6 @@ import { SynConfirmationDialog } from '../../components/dialogs/SynConfirmationD
 import { DisabledWithTooltip } from '../../components/DisabledWithTooltip'
 import { IconLabel } from '../../components/IconLabel'
 import type { KebabAction } from '../../components/SynKebabMenu'
-import { SynKebabMenu } from '../../components/SynKebabMenu'
 import { SynListPanelTable, SynListPanelToolbar, SynListPanelView } from '../../components/panels/list/SynListPanel'
 import { SynEmptyStateNoData } from '../../components/states/SynEmptyStateNoData'
 import { DateCell } from '../../components/table/DateCell'
