+++ frontend/packages/syntara-ui/src/routes/access-management/projects/ProjectRoleAssignmentsTab.tsx
@@ -10,7 +10,6 @@ import { DisabledWithTooltip } from '../../../components/DisabledWithTooltip'
 import { IconLabel } from '../../../components/IconLabel'
 import { SynLabel } from '../../../components/labels/SynLabel'
 import type { KebabAction } from '../../../components/SynKebabMenu'
-import { SynKebabMenu } from '../../../components/SynKebabMenu'
 import { SynListPanelTable, SynListPanelToolbar, SynListPanelView } from '../../../components/panels/list/SynListPanel'
 import { SynEmptyStateNoData } from '../../../components/states/SynEmptyStateNoData'
 import { invalidateAuthzCaches } from '../../../hooks/invalidateAuthzCaches'
